Change field class registration.
Since there does not seem to be a use case to register a different class
for each node. Instead of registering the field_class with each node and
passing it on to child nodes. We choose to use 'upward delegation' instead.
We register the class once in the root namespace. Other nodes further
down the tree will look up the ancestors for a field_class.

[#43]