Introduce a Factory to build nodes
This replaces the field_class argument and is created in the root namespace
as default argument. It is used in create_child to build the different node types.
The factory can delegate itself to the nodes it builds.

This allows for a custom factory to be passed to build field nodes with their custom
component implementations. Because of lexical reasons it does not resolve Field
class lookup as you might think when subclassing the factory. So we look for a Field
class in the current namespace, or fallback to Superform::Field.

This way it also made more sense for Field to not be a innerclass to Form. This introduced an unnecessary class that isn't needed when dealing with Superform
basics of node management. And also enables further splitting of Rails namespace
in seperate files.

Registration of a node now raises a custom error when a field was already
registered by that key.